+++ pom.xml
@@ -36,6 +36,7 @@
 		<module>spring-shell-docs</module>
 		<module>spring-shell-dependencies</module>
 		<module>spring-shell-starter</module>
+		<module>spring-shell-autoconfigure</module>
 	</modules>
 
 	<dependencyManagement>
@@ -73,6 +74,11 @@
 			<dependency>
 				<groupId>org.springframework.shell</groupId>
 				<artifactId>spring-shell-table</artifactId>
+				<version>3.0.0-SNAPSHOT</version>
+			</dependency>
+			<dependency>
+				<groupId>org.springframework.shell</groupId>
+				<artifactId>spring-shell-autoconfigure</artifactId>
 				<version>3.0.0-SNAPSHOT</version>
 			</dependency>
 			<dependency>
