+++ spring-shell-autoconfigure/src/main/java/org/springframework/shell/boot/JCommanderParameterResolverAutoConfiguration.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017 the original author or authors.
+ * Copyright 2017-2021 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,12 +14,13 @@
  * limitations under the License.
  */
 
-package org.springframework.shell.jcommander;
+package org.springframework.shell.boot;
 
 import com.beust.jcommander.JCommander;
 
 import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.shell.jcommander.JCommanderParameterResolver;
 import org.springframework.context.annotation.Bean;
 
 /**
@@ -28,7 +29,7 @@ import org.springframework.context.annotation.Bean;
  * @author Eric Bottard
  */
 @Configuration
-@ConditionalOnClass(JCommander.class)
+@ConditionalOnClass({ JCommander.class, JCommanderParameterResolver.class })
 public class JCommanderParameterResolverAutoConfiguration {
 
 	@Bean
