+++ spring-shell-autoconfigure/src/main/java/org/springframework/shell/boot/JLineAutoConfiguration.java
@@ -13,22 +13,25 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.springframework.shell;
+package org.springframework.shell.boot;
 
-import org.springframework.beans.factory.ObjectProvider;
+import org.jline.reader.impl.history.DefaultHistory;
+
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
-@Configuration(proxyBeanMethods = false)
-public class CommandRegistryAutoConfiguration {
+@Configuration
+public class JLineAutoConfiguration {
 
-	@Bean
-	public CommandRegistry commandRegistry(
-			ObjectProvider<MethodTargetRegistrar> methodTargerRegistrars) {
-		ConfigurableCommandRegistry registry = new ConfigurableCommandRegistry();
-		methodTargerRegistrars.orderedStream().forEach(resolver -> {
-			resolver.register(registry);
-		});
-		return registry;
+	@Configuration
+	@ConditionalOnMissingBean(org.jline.reader.History.class)
+	public static class JLineHistoryConfiguration {
+
+		@Bean
+		public org.jline.reader.History history(@Value("${spring.application.name:spring-shell}.log") String historyPath) {
+			return new DefaultHistory();
+		}
 	}
 }
