+++ spring-shell-autoconfigure/src/main/resources/META-INF/spring.factories
@@ -1,0 +1,10 @@
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
+org.springframework.shell.boot.SpringShellAutoConfiguration,\
+org.springframework.shell.boot.ApplicationRunnerAutoConfiguration,\
+org.springframework.shell.boot.CommandRegistryAutoConfiguration,\
+org.springframework.shell.boot.LineReaderAutoConfiguration,\
+org.springframework.shell.boot.CompleterAutoConfiguration,\
+org.springframework.shell.boot.JLineAutoConfiguration,\
+org.springframework.shell.boot.JLineShellAutoConfiguration,\
+org.springframework.shell.boot.JCommanderParameterResolverAutoConfiguration,\
+org.springframework.shell.boot.StandardCommandsAutoConfiguration
