+++ spring-shell-core/src/main/java/org/springframework/shell/jline/ExtendedDefaultParser.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017 the original author or authors.
+ * Copyright 2017-2021 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -35,7 +35,7 @@ import org.springframework.shell.CompletingParsedLine;
  * @author Original JLine author
  * @author Eric Bottard
  */
-class ExtendedDefaultParser implements Parser {
+public class ExtendedDefaultParser implements Parser {
 
 	private char[] quoteChars = { '\'', '"' };
