+++ spring-shell-docs/pom.xml
@@ -4,7 +4,7 @@
 
 	<artifactId>spring-shell-docs</artifactId>
 	<name>Spring Shell Documentation</name>
-	<packaging>jar</packaging>
+	<packaging>pom</packaging>
 
 	<parent>
 		<groupId>org.springframework.shell</groupId>
