--- spring-shell-standard-commands/src/main/resources/META-INF/spring.factories
@@ -1,2 +1,0 @@
-org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-org.springframework.shell.standard.commands.StandardCommandsAutoConfiguration
